create lambda function for sending emails
