--- CQCLocationAPICall/.gitignore
@@ -1,7 +1,0 @@
-# package directories
-node_modules
-jspm_packages
-
-# Serverless directories
-.serverless
-pcodefile
