--- CQCLocationAPICall/cqc-location-call-_star_-eu-west-1-policy.json
@@ -1,205 +1,0 @@
-{
-  "Version": "2012-10-17",
-  "Statement": [
-    {
-      "Effect": "Allow",
-      "Action": [
-        "cloudformation:List*",
-        "cloudformation:Get*",
-        "cloudformation:ValidateTemplate"
-      ],
-      "Resource": [
-        "*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "cloudformation:CreateStack",
-        "cloudformation:CreateUploadBucket",
-        "cloudformation:DeleteStack",
-        "cloudformation:Describe*",
-        "cloudformation:UpdateStack"
-      ],
-      "Resource": [
-        "arn:aws:cloudformation:eu-west-1:*:stack/cqc-location-call-*/*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "lambda:Get*",
-        "lambda:List*",
-        "lambda:CreateFunction",
-        "lambda:UpdateFunctionCode",
-        "lambda:UpdateFunctionConfiguration",
-        "lambda:PublishVersion"
-      ],
-      "Resource": [
-        "*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "s3:GetBucketLocation",
-        "s3:CreateBucket",
-        "s3:DeleteBucket",
-        "s3:ListBucket",
-        "s3:ListBucketVersions",
-        "s3:PutAccelerateConfiguration",
-        "s3:GetEncryptionConfiguration",
-        "s3:PutEncryptionConfiguration"
-      ],
-      "Resource": [
-        "arn:aws:s3:::cqc-location-call*serverlessdeploy*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "s3:PutObject",
-        "s3:GetObject",
-        "s3:DeleteObject"
-      ],
-      "Resource": [
-        "arn:aws:s3:::cqc-location-call*serverlessdeploy*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "lambda:AddPermission",
-        "lambda:CreateAlias",
-        "lambda:DeleteFunction",
-        "lambda:InvokeFunction",
-        "lambda:PublishVersion",
-        "lambda:RemovePermission",
-        "lambda:Update*"
-      ],
-      "Resource": [
-        "arn:aws:lambda:eu-west-1:*:function:cqc-location-call-*-*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "apigateway:GET",
-        "apigateway:POST",
-        "apigateway:PUT",
-        "apigateway:DELETE",
-        "apigateway:PATCH"
-      ],
-      "Resource": [
-        "arn:aws:apigateway:*::/restapis*",
-        "arn:aws:apigateway:*::/apikeys*",
-        "arn:aws:apigateway:*::/usageplans*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "iam:PassRole"
-      ],
-      "Resource": [
-        "arn:aws:iam::*:role/*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": "kinesis:*",
-      "Resource": [
-        "arn:aws:kinesis:*:*:stream/cqc-location-call-*-eu-west-1"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "iam:GetRole",
-        "iam:CreateRole",
-        "iam:PutRolePolicy",
-        "iam:DeleteRolePolicy",
-        "iam:DeleteRole"
-      ],
-      "Resource": [
-        "arn:aws:iam::*:role/cqc-location-call-*-eu-west-1-lambdaRole"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": "sqs:*",
-      "Resource": [
-        "arn:aws:sqs:*:*:cqc-location-call-*-eu-west-1"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "cloudwatch:GetMetricStatistics"
-      ],
-      "Resource": [
-        "*"
-      ]
-    },
-    {
-      "Action": [
-        "logs:CreateLogGroup",
-        "logs:CreateLogStream",
-        "logs:DeleteLogGroup"
-      ],
-      "Resource": [
-        "arn:aws:logs:eu-west-1:*:*"
-      ],
-      "Effect": "Allow"
-    },
-    {
-      "Action": [
-        "logs:PutLogEvents"
-      ],
-      "Resource": [
-        "arn:aws:logs:eu-west-1:*:*"
-      ],
-      "Effect": "Allow"
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "logs:DescribeLogStreams",
-        "logs:DescribeLogGroups",
-        "logs:FilterLogEvents"
-      ],
-      "Resource": [
-        "*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "events:Put*",
-        "events:Remove*",
-        "events:Delete*"
-      ],
-      "Resource": [
-        "arn:aws:events:*:*:rule/cqc-location-call-*-eu-west-1"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "events:DescribeRule"
-      ],
-      "Resource": [
-        "arn:aws:events:eu-west-1:*:rule/cqc-location-call-*-*"
-      ]
-    },
-    {
-      "Effect": "Allow",
-      "Action": [
-        "s3:CreateBucket"
-      ],
-      "Resource": [
-        "arn:aws:s3:::*"
-      ]
-    }
-  ]
-}
