--- CQCLocationAPICall/local.js
@@ -1,3 +1,0 @@
-const handler = require('./changes').handler;
-
-handler();
