--- CQCLocationAPICall/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "cqc-location-call",
-  "version": "1.0.0",
-  "description": "Call CQC Location API and populate our own location database.",
-  "scripts": {
-    "test": "nyc --reporter=html mocha test/**/*.js"
-  },
-  "author": "",
-  "license": "ISC",
-  "keywords": [],
-  "devDependencies": {
-    "aws-sdk-mock": "^4.5.0",
-    "axios-mock-adapter": "^1.17.0",
-    "serverless-offline": "^5.12.0",
-    "serverless-offline-sqs": "^3.0.1",
-    "serverless-webpack": "^5.3.1"
-  },
-  "dependencies": {
-    "axios": "^0.18.0",
-    "axios-concurrency": "^1.0.3",
-    "axios-retry": "^3.1.2",
-    "camelcase": "^5.3.1",
-    "convict": "^5.2.0",
-    "events": "^3.0.0",
-    "fast-csv": "^4.3.1",
-    "js-yaml": "^3.13.1",
-    "pg": "^7.6.0",
-    "pg-hstore": "^2.3.2",
-    "sequelize": "^4.44.3",
-    "sequelize-cli": "^5.2.0",
-    "sequelize-mock": "^0.10.2"
-  }
-}
