--- CQCLocationAPICall/serverless.yml
@@ -1,89 +1,0 @@
-service: cqc-location-call-${self:provider.stage}
-
-provider:
-  name: aws
-  runtime: nodejs10.x
-  memorySize: 512
-  stage: ${opt:stage, 'dev'}
-  iamRoleStatements:
-    - Effect: "Allow"
-      Action:
-        - "s3:PutObject"
-        - "s3:GetObject"
-      Resource:
-        Fn::Join:
-          - ""
-          - - arn:aws:s3:::*
-
-    - Effect: "Allow"
-      Action:
-        - "sqs:sendMessage"
-        - "sqs:recieveMessage"
-        - "sqs:deleteMessage"
-      Resource:
-        Fn::Join:
-          - ""
-          - - ${self:custom.sqsQueueArn.${self:provider.stage}}
-    - Effect: "Allow"
-      Action:
-        - "secretsmanager:getSecretValue"
-      Resource:
-        Fn::Join:
-          - ""
-          - - arn:aws:secretsmanager:*:*:*
-  region: eu-west-1
-  environment:
-    CQC_SQS_QUEUE: ${self:custom.sqsQueueUrl.${self:provider.stage}}
-    CQC_WALLET_ID: ${self:custom.wallet.${self:provider.stage}}
-    CQC_BUCKET: ${self:custom.bucket.${self:provider.stage}}
-    NODE_ENV: ${self:custom.nodeenv.${self:provider.stage}}
-
-include:
-  - config/*
-  - models/*
-  - aws/*
-  - changes.js
-  - update.js
-
-functions:
-  changes:
-    handler: changes.handler
-    name: cqc-changes-${self:provider.stage}
-    timeout: 900
-    memorySize: 1536
-    events:
-      - schedule: rate(24 hours)
-  update:
-    handler: update.handler
-    name: cqc-update-${self:provider.stage}
-    timeout: 900
-    reservedConcurrency: 1
-    events:
-      - sqs:
-        arn: ${self:custom.sqsQueueArn.${self:provider.stage}}
-        batchSize: 1289975
-
-# plugins:
-#   - serverless-offline
-
-custom:
-  sqsQueueArn:
-    dev: arn:aws:sqs:eu-west-1:364648107127:CQC_SQS_Dev
-    preprod: arn:aws:sqs:eu-west-1:364648107127:CQC_SQS_Preprod
-    prod: arn:aws:sqs:eu-west-1:364648107127:CQC_SQS_Prod
-  sqsQueueUrl:
-    dev: https://sqs.eu-west-1.amazonaws.com/364648107127/CQC_SQS_Dev
-    preprod: https://sqs.eu-west-1.amazonaws.com/364648107127/CQC_SQS_Preprod
-    prod: https://sqs.eu-west-1.amazonaws.com/364648107127/CQC_SQS_Prod
-  wallet:
-    dev: arn:aws:secretsmanager:eu-west-1:364648107127:secret:dev/api-JSfSOr
-    preprod: arn:aws:secretsmanager:eu-west-1:364648107127:secret:preprod/api-8ARd5F
-    prod: arn:aws:secretsmanager:eu-west-1:364648107127:secret:prod/api-KUX494
-  bucket:
-    dev: cqcs3dev
-    preprod: cqcs3preprod
-    prod: cqcs3prod
-  nodeenv:
-    dev: development
-    preprod: production
-    prod: production
