--- CQCLocationAPICall/test/mockdata/s3.json
@@ -1,1 +1,0 @@
-{}
