+++ lambdas/CQCLocationAPICall/.gitignore
@@ -1,0 +1,7 @@
+# package directories
+node_modules
+jspm_packages
+
+# Serverless directories
+.serverless
+pcodefile
