+++ lambdas/CQCLocationAPICall/config/development.yaml
@@ -1,0 +1,11 @@
+aws:
+  region: eu-west-1
+  secrets:
+    use: true
+log:
+  sequelize: false
+db:
+  ssl: true
+  client_ssl:
+    status: true
+    usingFiles: false
