+++ lambdas/CQCLocationAPICall/config/localhost.yaml
@@ -1,0 +1,13 @@
+aws:
+  region: eu-west-1
+  bucketname: cqcs3lambda
+  sqsqueue: https://sqs.eu-west-1.amazonaws.com/364648107127/LamdaCQC
+  secrets:
+    use: false
+log:
+  sequelize: true
+db:
+  ssl: false
+  client_ssl:
+    status: false
+    usingFiles: false
