+++ lambdas/CQCLocationAPICall/event.json
@@ -1,0 +1,9 @@
+{"Records": [
+    {
+      "messageId": "ca1dfc31-8e3d-4121-8e72-bb35265039dc",
+      "receiptHandle": "AQEBHgBxB1qn0uH1rjfAyJIw4aF/i/60tKgkwWJYFh1qCipAp1UTrMvJkM9nkyD4wBo1uFzal50h92kc9lBmvTIJW5oyi+jU2DDIEuoAm551l6BIQwnmdZOn7xVWTnzr19LRu58AtW4t8k2DqNw6ETHMQFfY3WC55VHjQkMr9Gr0Hg8sn5+0uQ99znOOWnrvnf0PTxYREl+97chJ2CadJG1Ui/kp32w9QXnUFmvfgF6NzlRrdvLRJVdL0gbCoo2zfw+2WBZ6GF3a0G0SaQFKRvVg8rxi+De7wBXBnH+3lkPL8CWh+tnGTJp7QqMHdKF5viKG1ABgjITv/jDArUyd84wLW2J03AqqfVN5wfs69cvDETl77V9Oyrt+A7UF+qNl7/G5",
+      "mD5OfBody": "121475819da76d0f0245065c61013f1a",
+      "body":"{\"locationId\":\"1-125856395\",\"status\":\"\",\"startDate\":\"2019-10-30T18:20:10Z\",\"endDate\":\"2019-10-31T15:06:47Z\"}"
+    }
+  ]
+}
