+++ lambdas/CQCLocationAPICall/local.js
@@ -1,0 +1,3 @@
+const handler = require('./changes').handler;
+
+handler();
