+++ lambdas/CQCLocationAPICall/test/mockdata/s3.json
@@ -1,0 +1,1 @@
+{}
